+++ renderer/units/Unicode Foldings/import.js
@@ -22,7 +22,9 @@ let defaultFolderPath;
 //
 module.exports.start = function (context)
 {
-    const { app } = require ('electron').remote;
+    const { remote } = require ('electron');
+    const { app } = remote;
+    //
     let defaultLocale = app.getLocale ();
     //
     const path = require ('path');