+++ renderer/units/Unihan Variants/import.js
@@ -1,6 +1,7 @@
 //
 const unit = document.getElementById ('unihan-variants-unit');
 //
+const historyButton = unit.querySelector ('.history-button');
 const unihanInput = unit.querySelector ('.unihan-input');
 const lookupButton = unit.querySelector ('.lookup-button');
 const extraVariantsCheckbox = unit.querySelector ('.extra-variants-checkbox');
@@ -17,7 +18,7 @@ const instructions = unit.querySelector ('.instructions');
 const references = unit.querySelector ('.references');
 const links = unit.querySelector ('.links');
 //
-const unihanHistorySize = 256;   // 0: unlimited
+const unihanHistorySize = 128;   // 0: unlimited
 //
 let unihanHistory = [ ];
 let unihanHistoryIndex = -1;
@@ -29,12 +30,14 @@ let defaultFolderPath;
 //
 module.exports.start = function (context)
 {
-    const { remote } = require ('electron');
+    const { remote, shell } = require ('electron');
+    const { Menu } = remote;
     //
     const fs = require ('fs');
     const path = require ('path');
     //
     const fileDialogs = require ('../../lib/file-dialogs.js');
+    const pullDownMenus = require ('../../lib/pull-down-menus.js');
     const linksList = require ('../../lib/links-list.js');
     //
     // https://github.com/mdaines/viz.js/wiki/Usage
@@ -92,7 +95,7 @@ module.exports.start = function (context)
         }
         return character;
     }
-    // 
+    //
     unihanInput.addEventListener
     (
         'input',
@@ -547,6 +550,7 @@ module.exports.start = function (context)
     function updateUnihanData (character)
     {
         unihanInput.value = "";
+        unihanInput.blur ();
         unihanInput.dispatchEvent (new Event ('input'));
         displayData (character);
     }
@@ -583,7 +587,7 @@ module.exports.start = function (context)
                 }
                 else
                 {
-                    remote.shell.beep ();
+                    shell.beep ();
                 }
             }
             else
@@ -592,6 +596,41 @@ module.exports.start = function (context)
                 unihanHistorySave = null;
                 updateUnihanData ("");
             }
+        }
+    );
+    //
+    function insertUnihanCharacter (menuItem)
+    {
+        unihanInput.value = menuItem.id;
+        unihanInput.dispatchEvent (new Event ('input'));
+        lookupButton.click ();
+    };
+    historyButton.addEventListener
+    (
+        'click',
+        (event) =>
+        {
+            let historyMenuTemplate = [ ];
+            if (unihanHistory.length > 0)
+            {
+                for (let unihan of unihanHistory)
+                {
+                    historyMenuTemplate.push
+                    (
+                        {
+                            label: `${unihan}${(process.platform === 'darwin') ? "\t" : "\xA0\xA0"}${unicode.characterToCodePoint (unihan)}`,
+                            id: unihan,
+                            click: insertUnihanCharacter
+                        }
+                    );
+                }
+            }
+            else
+            {
+                historyMenuTemplate.push ({ label: "(no history yet)", enabled: false });
+            }
+            let historyContextualMenu = Menu.buildFromTemplate (historyMenuTemplate);
+            pullDownMenus.popup (event.currentTarget, historyContextualMenu);
         }
     );
     //