+++ CHANGES.md
@@ -2,6 +2,15 @@
 
 This project adheres to [Semantic Versioning](https://semver.org/).
 
+## 9.1.0
+
+- Added history pop-up menus for specimens and looked up characters to the **Unicode Data Finder**, **Unihan Components**, **Unihan Data Finder**, **Unihan Inspector**, and **Unihan Variants** utilities.
+- Used dimmed style effect for nested-matching characters displayed in the **Match IDS** feature of the **Unihan Components** utility.
+- Fixed issue in the **Match IDS** feature of the **Unihan Components** utility, where some characters would not be nested-matched by mistake in specific cases.
+- Updated `IDS.TXT` data file and `BabelStoneHanPUA.woff2` font file.
+- Improved styling of <code> and <kbd> tags.
+- Added `Confirm Quit` dialog.
+
 ## 9.0.0
 
 - Added new experimental **Unihan Components** utility, made up of two features: **Lookup IDS** and **Match IDS**.